+++ aui.views/src/AUI/Font/AFont.cpp
@@ -20,12 +20,6 @@
 #include "AUI/Common/AStringVector.h"
 #include "AFont.h"
 
-#if AUI_PLATFORM_LINUX
-#include <fontconfig/fontconfig.h>
-#elif AUI_PLATFORM_WIN
-#include <windows.h>
-#endif
-
 
 AFont::AFont(AFontManager* fm, const AString& path) :
         ft(fm->mFreeType) {
@@ -72,168 +66,6 @@ bool AFont::isItalic() const {
 
 bool AFont::hasGlyph(char32_t codepoint) const {
     return FT_Get_Char_Index(mFace, codepoint) != 0;
-}
-
-void AFontManager::ensureFallbackFace() {
-    if (mFallbackFace) {
-        return; // already loaded
-    }
-    if (mFallbackAttempted) {
-        // Already tried and failed; don't retry every glyph
-        return;
-    }
-    mFallbackAttempted = true;
-
-#if AUI_PLATFORM_LINUX
-    // Use fontconfig to find a CJK-capable sans-serif font.
-    // Ensure fontconfig is initialized before using default config (nullptr).
-    if (!FcInit()) {
-        ALogger::warn("Font") << "FcInit() failed; CJK fallback unavailable";
-        return;
-    }
-    // Use a charset-based query with a common CJK character to ensure the font actually has CJK glyphs.
-    FcPattern* pattern = FcPatternCreate();
-    FcCharSet* charset = FcCharSetCreate();
-    FcCharSetAddChar(charset, 0x4E2D); // U+4E2D = '中' (common CJK character)
-    FcPatternAddCharSet(pattern, FC_CHARSET, charset);
-    FcPatternAddString(pattern, FC_FAMILY, reinterpret_cast<const FcChar8*>("sans"));
-    FcConfigSubstitute(nullptr, pattern, FcMatchPattern);
-    FcDefaultSubstitute(pattern);
-    FcCharSetDestroy(charset);
-
-    FcResult result;
-    FcPattern* match = FcFontMatch(nullptr, pattern, &result);
-    if (match) {
-        if (result == FcResultMatch) {
-            FcChar8* path = nullptr;
-            if (FcPatternGetString(match, FC_FILE, 0, &path) == FcResultMatch) {
-                mFallbackFontPath = reinterpret_cast<const char*>(path);
-
-                // Try to load the fallback font face.
-                // TTC collections may need face_index > 0; try index 0 first (usually Regular).
-                if (FT_New_Face(mFreeType->getFt(), mFallbackFontPath.toStdString().c_str(), 0, &mFallbackFace)) {
-                    ALogger::warn("Font") << "Failed to load fallback font: " << mFallbackFontPath;
-                    mFallbackFace = nullptr;
-                    mFallbackFontPath.clear();
-                } else {
-                    ALogger::info("Font") << "Loaded CJK fallback font: " << mFallbackFontPath;
-                }
-            }
-        }
-        FcPatternDestroy(match);
-    }
-    FcPatternDestroy(pattern);
-#elif AUI_PLATFORM_WIN
-    // Try known CJK font files from the Windows Fonts directory.
-    // Order: most-comprehensive first (Microsoft YaHei covers CJK + Kana,
-    // Malgun Gothic covers CJK + Hangul, then region-specific fallbacks).
-    const char* cjkFonts[] = {
-        "msyh.ttc",       // Microsoft YaHei (Simplified Chinese, includes CJK + Kana)
-        "malgun.ttf",     // Malgun Gothic (Korean, includes Hangul + CJK)
-        "simsun.ttc",     // SimSun (Chinese Traditional)
-        "msgothic.ttc",   // MS Gothic (Japanese)
-        "yugothic.ttf",   // Yu Gothic (Japanese)
-        "meiryo.ttc",     // Meiryo (Japanese)
-    };
-    static const AString fontsDir = [] {
-        wchar_t buf[MAX_PATH];
-        UINT len = GetWindowsDirectoryW(buf, MAX_PATH);
-        if (len > 0 && len < MAX_PATH) {
-            return AString(reinterpret_cast<char16_t*>(buf)) + "\\Fonts\\";
-        }
-        ALogger::warn("Font") << "GetWindowsDirectoryW() failed, falling back to C:\\Windows\\Fonts\\";
-        return AString("C:\\Windows\\Fonts\\");
-    }();
-    for (auto& f : cjkFonts) {
-        AString fontPath = fontsDir + f;
-        if (FT_New_Face(mFreeType->getFt(), fontPath.toStdString().c_str(), 0, &mFallbackFace) == 0) {
-            mFallbackFontPath = fontPath;
-            ALogger::info("Font") << "Loaded CJK fallback font: " << mFallbackFontPath;
-            break;
-        }
-    }
-    if (!mFallbackFace) {
-        ALogger::warn("Font") << "No CJK fallback font found on Windows";
-    }
-#elif AUI_PLATFORM_MACOS
-    // macOS system CJK fonts (stable paths since OS X 10.11).
-    const char* cjkFontsMac[] = {
-        "/System/Library/Fonts/PingFang.ttc",
-        "/System/Library/Fonts/AppleSDGothicNeo.ttc",
-        "/System/Library/Fonts/Hiragino Sans.ttc",
-        "/System/Library/Fonts/Supplemental/AppleSDGothicNeo.ttc",
-    };
-    for (auto& f : cjkFontsMac) {
-        AString fontPath(f);
-        if (FT_New_Face(mFreeType->getFt(), fontPath.toStdString().c_str(), 0, &mFallbackFace) == 0) {
-            mFallbackFontPath = fontPath;
-            ALogger::info("Font") << "Loaded CJK fallback font: " << mFallbackFontPath;
-            break;
-        }
-    }
-    if (!mFallbackFace) {
-        ALogger::warn("Font") << "No CJK fallback font found on macOS";
-    }
-#elif AUI_PLATFORM_ANDROID
-    // Android system CJK fonts (varies by API level).
-    const char* cjkFontsAndroid[] = {
-        "/system/fonts/NotoSansCJK-Regular.ttc",  // Android 5-9
-        "/system/fonts/NotoSansSC-Regular.otf",    // Android 10+ (Chinese)
-        "/system/fonts/NotoSansKR-Regular.otf",    // Android 10+ (Korean)
-        "/system/fonts/NotoSansJP-Regular.otf",    // Android 10+ (Japanese)
-    };
-    for (auto& f : cjkFontsAndroid) {
-        AString fontPath(f);
-        if (FT_New_Face(mFreeType->getFt(), fontPath.toStdString().c_str(), 0, &mFallbackFace) == 0) {
-            mFallbackFontPath = fontPath;
-            ALogger::info("Font") << "Loaded CJK fallback font: " << mFallbackFontPath;
-            break;
-        }
-    }
-    if (!mFallbackFace) {
-        ALogger::warn("Font") << "No CJK fallback font found on Android";
-    }
-#elif AUI_PLATFORM_IOS
-    // iOS: try system font paths first (some iOS versions allow reading them),
-    // then fall back to a bundled resource if available.
-    const char* cjkFontsIos[] = {
-        "/System/Library/Fonts/PingFang.ttc",
-        "/System/Library/Fonts/AppleSDGothicNeo.ttc",
-    };
-    {
-        bool loaded = false;
-        for (auto& f : cjkFontsIos) {
-            AString fontPath(f);
-            if (FT_New_Face(mFreeType->getFt(), fontPath.toStdString().c_str(), 0, &mFallbackFace) == 0) {
-                mFallbackFontPath = fontPath;
-                ALogger::info("Font") << "Loaded CJK fallback font: " << mFallbackFontPath;
-                loaded = true;
-                break;
-            }
-        }
-        if (!loaded) {
-            // Try bundled resource
-            try {
-                mFallbackFontDataBuffer = AByteBuffer::fromStream(AUrl(":uni/font/NotoSansCJK-Fallback.ttf").open());
-                if (FT_New_Memory_Face(mFreeType->getFt(),
-                        (const FT_Byte*) mFallbackFontDataBuffer.data(),
-                        mFallbackFontDataBuffer.getSize(), 0, &mFallbackFace) == 0) {
-                    mFallbackFontPath = AString(":uni/font/NotoSansCJK-Fallback.ttf");
-                    ALogger::info("Font") << "Loaded CJK fallback font from bundle";
-                    loaded = true;
-                }
-            } catch (...) {
-                // bundled font not present
-            }
-        }
-        if (!loaded) {
-            ALogger::warn("Font") << "No CJK fallback font found on iOS";
-        }
-    }
-#else
-    // Unknown platform; mark fallback as unavailable.
-    ALogger::warn("Font") << "CJK font fallback not available on this platform";
-#endif
 }
 
 glm::vec2 AFont::getKerning(wchar_t left, wchar_t right) {
